perf(tests): task-1462 — Chatbooks shares the ChaChaNotes template; task-1461 Media_DB measured no-go
Only CharactersRAGDB's DDL is worth templating (137ms vs Prompts 6.8ms /
Media 10.6ms — Media_DB's whole directory already runs in 3.26s, so task-1461
closes as a measured no-go with no code change). The session template fixture
moves to the root conftest (lazy import; resolution walks up the chain, so
task-1460's call sites work unchanged) and Chatbooks' four
CharactersRAGDB-constructing fixtures seed from it at the path seam — zero
test-body churn. Mid-test import-side constructions keep building fresh: a
fresh import target is the semantics under test. Chatbooks 23.99s -> 20.78s,
171 passed unchanged; ChaChaNotesDB re-verified 4.29s post-hoist.
